无

+++ Visual Studio/C#/Eruru.Json For Shared Project/JsonValueType.cs
@@ -4,13 +4,13 @@
 
 		Null = 1 << 0,
 		Decimal = 1 << 1,
-		Long = 1 << 2,
+		Integer = 1 << 2,
 		Bool = 1 << 3,
 		String = 1 << 4,
 		DateTime = 1 << 5,
 		Array = 1 << 6,
 		Object = 1 << 7,
-		Value = Null | Decimal | Long | Bool | String | DateTime
+		Value = Null | Decimal | Integer | Bool | String | DateTime
 
 	}
